Refactor: 예외 처리 로직을 `global/exception`로 이동
