+++ src/main/java/com/team7/retriever/global/exception/BadRequestException.java
@@ -1,6 +1,6 @@
-package com.team7.retriever.exception;
+package com.team7.retriever.global.exception;
 
-import com.team7.retriever.exception.base.BaseErrorCode;
+import com.team7.retriever.global.exception.base.BaseErrorCode;
 
 public class BadRequestException extends CustomException {
 	public BadRequestException(final BaseErrorCode baseErrorCode) {
