+++ src/main/java/com/team7/retriever/global/exception/ConflictException.java
@@ -1,6 +1,6 @@
-package com.team7.retriever.exception;
+package com.team7.retriever.global.exception;
 
-import com.team7.retriever.exception.base.BaseErrorCode;
+import com.team7.retriever.global.exception.base.BaseErrorCode;
 
 public class ConflictException extends CustomException {
 	public ConflictException(final BaseErrorCode baseErrorCode) {
