+++ src/main/java/com/team7/retriever/global/exception/NotFoundException.java
@@ -1,6 +1,6 @@
-package com.team7.retriever.exception;
+package com.team7.retriever.global.exception;
 
-import com.team7.retriever.exception.base.BaseErrorCode;
+import com.team7.retriever.global.exception.base.BaseErrorCode;
 
 public class NotFoundException extends CustomException {
 	public NotFoundException(final BaseErrorCode baseErrorCode) {
