+++ src/main/java/com/team7/retriever/global/exception/base/BaseErrorCode.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.exception.base;
+package com.team7.retriever.global.exception.base;
 
 public interface BaseErrorCode {
 	int getStatus();
