+++ src/main/java/com/team7/retriever/service/AuthService.java
@@ -25,11 +25,11 @@ import com.team7.retriever.dto.response.LoginSuccessResponse;
 import com.team7.retriever.dto.response.ReissueResponse;
 import com.team7.retriever.entity.User;
 import com.team7.retriever.entity.enums.Role;
-import com.team7.retriever.exception.BadRequestException;
-import com.team7.retriever.exception.CustomException;
-import com.team7.retriever.exception.NotFoundException;
-import com.team7.retriever.exception.UnauthorizedException;
-import com.team7.retriever.exception.UserErrorCode;
+import com.team7.retriever.global.exception.BadRequestException;
+import com.team7.retriever.global.exception.CustomException;
+import com.team7.retriever.global.exception.NotFoundException;
+import com.team7.retriever.global.exception.UnauthorizedException;
+import com.team7.retriever.global.exception.UserErrorCode;
 import com.team7.retriever.repository.UserRepository;
 
 @Slf4j
